pyproject.toml
[tool.poetry]
name = "birdset"
version = "0.2.0"
description = "BirdSet: A Large-Scale Dataset for Audio Classification in Avian Bioacoustics"
authors = ["Lukas Rauch <lukas.rauch@uni-kassel.de>", "Raphael Schwinger <rsc@informatik.uni-kiel.de>", "Moritz Wirth <moritz.wirt@uni-kassel.de>", "Rene  Heinrich <rene.heinrich@iee.fraunhofer.de>", "Jonas Lange <jla@informatik.uni-kiel.de>"]
readme = "README.md"
packages = [
    { include = "ajepa", from = "projects/" },
    { include = "birdset", from = "./" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
scipy = "^1.11.3"
ffprobe = "^0.5"
# --------- pytorch --------- #
torch = "2.2.0"
torchvision = "0.17.0"

lightning= "^2.0"
torchmetrics= "1.2.0"

# --------- hydra --------- #
hydra-core = "1.3.2"
hydra-colorlog = "1.2"

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
# rootutils       # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
# rich            # beautiful text formatting in terminal
# pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
rootutils = "^1.0.7"
pre-commit = "^3.5.0"
rich = "13.7.0"
matplotlib = "^3.8.0"
tensorboard = "^2.14.1"
ipywidgets = "^8.1.1"
jupyter = "^1.0.0"
datasets = "^3.3.1"
audiomentations = "0.34.1"
torch-audiomentations = "0.11.0"
wandb = "^0.16.0"
transformers = "4.38.0"
pyrootutils = "^1.0.4"
mlflow = "^2.11.0"
pydub = "^0.25.1"
ipython = "8.18.1"
librosa = "0.10.1"
numpy = "1.26.4"
torchaudio = "2.2.0"
scikit-learn = "1.3.2"
pillow = "10.1.0"
pandas = "2.1.3"
soundfile = "0.12.1"
black = {extras = ["jupyter"], version = "25.1.0"}
pytest-timeout = "^2.3.1"
pytest-html = "^4.1.1"
huggingface-hub = "^0.28.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
